dedupe golden fixture builders, drop orphaned python golden files
