--- python/tests/golden/interactive_exec_result_error.json
@@ -1,9 +1,0 @@
-{
-  "error": "CommandBlocked: 'exit'",
-  "output": "",
-  "session_id": "sess-0001",
-  "timestamp": "2026-01-01T00:00:00",
-  "execution_time": 0.0,
-  "command_count": 0,
-  "buffer_size": 0
-}
