--- python/tests/golden/interactive_exec_result_success.json
@@ -1,9 +1,0 @@
-{
-  "error": null,
-  "output": "OpenROAD v2.0",
-  "session_id": "sess-0001",
-  "timestamp": "2026-01-01T00:00:00",
-  "execution_time": 1.5,
-  "command_count": 3,
-  "buffer_size": 2048
-}
