--- python/tests/golden/interactive_session_info_error.json
@@ -1,10 +1,0 @@
-{
-  "error": "Session failed to start",
-  "session_id": "sess-0002",
-  "created_at": "2026-01-01T00:00:00",
-  "is_alive": false,
-  "command_count": 0,
-  "buffer_size": 0,
-  "uptime_seconds": null,
-  "state": null
-}
