--- python/tests/golden/interactive_session_info_success.json
@@ -1,10 +1,0 @@
-{
-  "error": null,
-  "session_id": "sess-0001",
-  "created_at": "2026-01-01T00:00:00",
-  "is_alive": true,
-  "command_count": 5,
-  "buffer_size": 4096,
-  "uptime_seconds": 12.5,
-  "state": "active"
-}
