--- python/tests/golden/interactive_session_list.json
@@ -1,27 +1,0 @@
-{
-  "error": null,
-  "sessions": [
-    {
-      "error": null,
-      "session_id": "sess-0001",
-      "created_at": "2026-01-01T00:00:00",
-      "is_alive": true,
-      "command_count": 5,
-      "buffer_size": 4096,
-      "uptime_seconds": 12.5,
-      "state": "active"
-    },
-    {
-      "error": "Session failed to start",
-      "session_id": "sess-0002",
-      "created_at": "2026-01-01T00:00:00",
-      "is_alive": false,
-      "command_count": 0,
-      "buffer_size": 0,
-      "uptime_seconds": null,
-      "state": null
-    }
-  ],
-  "total_count": 2,
-  "active_count": 1
-}
