--- python/tests/golden/list_images.json
@@ -1,26 +1,0 @@
-{
-  "error": null,
-  "run_path": "/reports/nangate45/gcd/base",
-  "total_images": 2,
-  "images_by_stage": {
-    "floorplan": [
-      {
-        "filename": "floorplan.webp",
-        "path": "/reports/nangate45/gcd/base/floorplan.webp",
-        "size_bytes": 15000,
-        "modified_time": "2026-01-01T00:00:00",
-        "type": "floorplan"
-      }
-    ],
-    "route": [
-      {
-        "filename": "route.webp",
-        "path": "/reports/nangate45/gcd/base/route.webp",
-        "size_bytes": 22000,
-        "modified_time": "2026-01-01T00:00:00",
-        "type": "route"
-      }
-    ]
-  },
-  "message": null
-}
