--- python/tests/golden/list_images_empty.json
@@ -1,7 +1,0 @@
-{
-  "error": null,
-  "run_path": null,
-  "total_images": null,
-  "images_by_stage": null,
-  "message": "No images found"
-}
