--- python/tests/golden/read_image.json
@@ -1,20 +1,0 @@
-{
-  "error": null,
-  "image_data": "aGVsbG8=",
-  "metadata": {
-    "filename": "floorplan.webp",
-    "format": "webp",
-    "size_bytes": 15000,
-    "width": 1024,
-    "height": 768,
-    "modified_time": "2026-01-01T00:00:00",
-    "stage": "floorplan",
-    "type": "floorplan",
-    "compression_applied": true,
-    "original_size_bytes": 48000,
-    "original_width": 2048,
-    "original_height": 1536,
-    "compression_ratio": 0.3125
-  },
-  "message": null
-}
