--- python/tests/golden/read_image_error.json
@@ -1,6 +1,0 @@
-{
-  "error": "Image not found: foo.png",
-  "image_data": null,
-  "metadata": null,
-  "message": null
-}
