--- python/tests/golden/session_termination.json
@@ -1,7 +1,0 @@
-{
-  "error": null,
-  "session_id": "sess-0001",
-  "terminated": true,
-  "was_alive": true,
-  "force": false
-}
