--- python/tests/golden/tool_manifest.json
@@ -1,196 +1,0 @@
-{
-  "create_interactive_session": {
-    "params": {
-      "session_id": {
-        "type": "string",
-        "required": false
-      },
-      "command": {
-        "type": "array",
-        "required": false
-      },
-      "env": {
-        "type": "object",
-        "required": false
-      },
-      "cwd": {
-        "type": "string",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": false,
-      "destructiveHint": false,
-      "idempotentHint": false,
-      "openWorldHint": false
-    }
-  },
-  "get_session_history": {
-    "params": {
-      "session_id": {
-        "type": "string",
-        "required": true
-      },
-      "limit": {
-        "type": "integer",
-        "required": false
-      },
-      "search": {
-        "type": "string",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "get_session_metrics": {
-    "params": {},
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "inspect_interactive_session": {
-    "params": {
-      "session_id": {
-        "type": "string",
-        "required": true
-      }
-    },
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "interactive_openroad_exec": {
-    "params": {
-      "command": {
-        "type": "string",
-        "required": true
-      },
-      "session_id": {
-        "type": "string",
-        "required": false
-      },
-      "timeout_ms": {
-        "type": "integer",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": false,
-      "destructiveHint": true,
-      "idempotentHint": false,
-      "openWorldHint": false
-    }
-  },
-  "interactive_openroad_query": {
-    "params": {
-      "command": {
-        "type": "string",
-        "required": true
-      },
-      "session_id": {
-        "type": "string",
-        "required": false
-      },
-      "timeout_ms": {
-        "type": "integer",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": false,
-      "openWorldHint": false
-    }
-  },
-  "list_interactive_sessions": {
-    "params": {},
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "list_report_images": {
-    "params": {
-      "platform": {
-        "type": "string",
-        "required": true
-      },
-      "design": {
-        "type": "string",
-        "required": true
-      },
-      "run_slug": {
-        "type": "string",
-        "required": true
-      },
-      "stage": {
-        "type": "string",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "read_report_image": {
-    "params": {
-      "platform": {
-        "type": "string",
-        "required": true
-      },
-      "design": {
-        "type": "string",
-        "required": true
-      },
-      "run_slug": {
-        "type": "string",
-        "required": true
-      },
-      "image_name": {
-        "type": "string",
-        "required": true
-      }
-    },
-    "annotations": {
-      "readOnlyHint": true,
-      "destructiveHint": false,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  },
-  "terminate_interactive_session": {
-    "params": {
-      "session_id": {
-        "type": "string",
-        "required": true
-      },
-      "force": {
-        "type": "boolean",
-        "required": false
-      }
-    },
-    "annotations": {
-      "readOnlyHint": false,
-      "destructiveHint": true,
-      "idempotentHint": true,
-      "openWorldHint": false
-    }
-  }
-}
